+++ packages/truewire/src/truewire/generation/python/types/__init__.py
@@ -1,4 +1,4 @@
-from .schema import Type, InlineType, Ref, Literal, List, Tuple, Union, Variant, Dict, Record, Field
+from .schema import Type, InlineType, Scalar, Ref, Literal, List, Tuple, Union, Variant, Dict, Record, Field
 from .parser import Parser
 from .code import CodeGenerator, GeneratorFn, Code, Imports, Renderer
 from .normalize import Normalizer